requirements.txt
requests==2.32.3
python-dotenv==1.0.1 
python-docx==1.1.2
